# Optimize Action Handling by handling all possible actions within a reasonable frame rate.