--- a/buildSrc/src/main/java/kotlin-jvm.gradle.kts
+++ b/buildSrc/src/main/java/kotlin-jvm.gradle.kts
@@ -9,4 +9,18 @@ extensions.getByType(JavaPluginExtension::class).apply {
   targetCompatibility = JavaVersion.VERSION_1_8
 }
 
+tasks.withType<Test> {
+  project
+    .properties
+    .asSequence()
+    .filter { (key, value) ->
+      key.startsWith("workflow.runtime") && value != null
+    }
+    .forEach { (key, value) ->
+      // Add in a system property to the fork for the test.
+      systemProperty(key, value!!)
+    }
+}
+
+
 project.kotlinCommonSettings(bomConfigurationName = "implementation")