--- a/gradle.properties
+++ b/gradle.properties
@@ -25,3 +25,7 @@ POM_DEVELOPER_ID=square
 POM_DEVELOPER_NAME=Square, Inc.
 POM_DEVELOPER_URL=https://github.com/square/
 SONATYPE_STAGING_PROFILE=com.squareup
+
+# For KMP in the IDE, until we get to Kotlin version 1.6.20
+kotlin.mpp.enableGranularSourceSetsMetadata=true
+kotlin.native.enableDependencyPropagation=false