--- a/samples/compose-samples/build.gradle.kts
+++ b/samples/compose-samples/build.gradle.kts
@@ -3,6 +3,7 @@ plugins {
   `kotlin-android`
   `android-sample-app`
   `android-ui-tests`
+  `compose-ui-tests`
 }
 
 android {