--- a/samples/compose-samples/src/androidTest/java/com/squareup/sample/compose/hellocomposeworkflow/HelloComposeWorkflowTest.kt
+++ b/samples/compose-samples/src/androidTest/java/com/squareup/sample/compose/hellocomposeworkflow/HelloComposeWorkflowTest.kt
@@ -25,6 +25,8 @@ class HelloComposeWorkflowTest {
     .around(IdlingDispatcherRule)
 
   @Test fun togglesBetweenStates() {
+    composeRule.activityRule.scenario
+
     composeRule.onNodeWithText("Hello")
       .assertIsDisplayed()
       .performClick()