--- a/samples/tutorial/tutorial-1-complete/src/main/java/workflow/tutorial/TutorialActivity.kt
+++ b/samples/tutorial/tutorial-1-complete/src/main/java/workflow/tutorial/TutorialActivity.kt
@@ -37,7 +37,7 @@ class TutorialViewModel(savedState: SavedStateHandle) : ViewModel() {
     renderWorkflowIn(
       workflow = WelcomeWorkflow,
       scope = viewModelScope,
-      savedStateHandle = savedState
+      savedStateHandle = savedState,
     )
   }
 }