--- a/workflow-config/config-android/gradle.properties
+++ b/workflow-config/config-android/gradle.properties
@@ -1,0 +1,3 @@
+POM_ARTIFACT_ID=workflow-config-android
+POM_NAME=Workflow Runtime Android Configuration
+POM_PACKAGING=aar