--- a/workflow-config/config-android/src/main/AndroidManifest.xml
+++ b/workflow-config/config-android/src/main/AndroidManifest.xml
@@ -1,0 +1,1 @@
+<manifest package="com.squareup.workflow1.config"/>