We introduce this change via a `RuntimeConfig` sealed interface that we can optionally pass to `renderWorkflowIn` which is then forwarded to the `WorkflowRunner` which makes the actual decisions about how to run. Renaming `nextOutput` to `processActions` (a bit more general as its doing more than getting an Output in the OutputT sense), we also change its return type to a sealed interface which can handle `PropsUpdated` and `TimeoutForFrame`. This means we erase the `OutputT` from the internal implementation's signature but I have not seen any issues as a result of that as it is used internally or after we cast back to a `WorkflowOutput<T>`.

--- testing discussion:

With multiple runtime configurations we do the `RenderWorkflowInTest`s and `WorkflowRunnerTest`s once with both runtimes.

To do this we've built a simple ParameterizedTestRunner for kotlin common test code.

For all the other tests, we use a property (`-Pworkflow.runtime=timeout`) to override the baseline default runtime. For jvm unit tests, this will be pulled into the jvm fork running the tests and then the test runtime will use this runtime as specified. For Android tests, we have to build the app with this property and then run the integration tests against that app built with a particular runtime.

Note that because we introduce this frame timeout in order to process multiple actions, we provide a rendering only as fast as the frame rate times out. So when we run the tests with the modified runtime we need to account for this. This means if we expect a change in UI tests, we will need to wait for the frame timeout before verifying, which means we have to brute force Espresso a bit.

Things are more complicated for Compose because the Recomposer also has its own dispatcher, so we don't just nee to wait for the Workflow runtime's frame timeout, we also then need to wait for the Recomposer to finish well ... recomposing.  We don't want to use `rememberCoroutineScope` for our Workflow runtime because while that will tie it to the scope of the composition, we can't control the dispatcher in that scope easily in order to advance past the Workflow runtime's frame timeout.

Those changes remain so that the tests support both runtimes.

--- back to implementation discussion:

Note further, we are currently using a hard-coded frame timeout of 30ms because its somewhere between 60fps's 16ms and @pyricau's research for touch interaction at 100ms. In tuning this value it also produced the best reduction in render passes of the values I tried (10ms, 16ms, 30ms). Theoretically much better than this hard coded value will be to integrate a monotonic frame clock, which for Android can be tied to the Android Choreographer. I noticed this is what molecule wraps to time the emitted view model updates - we should take a look there and see if we can apply the same mechanism. See https://github.com/cashapp/molecule#frame-clock . Fine tuning this is covered by #810 .

We also introduce `currentTimeMillis()` expect so that we can calculate the timeout length on all platforms as we process actions.

Addresses #370 .