@steve-the-edwards 2022-06-08
Will still need a `currentTimeMillis` implementation for ios. cinterop or NSDate or kotlinx-datetime (this last one seems heavy?)

@bnvinay92 Do you have any good thoughts for this for `ios` target?