@rjrjr 2022-06-17
> We should keep running _all_ these tests for both runtimes, but I haven't thought of a good way to to do that yet and it's a lot of work to set up the test matrix for runtime versions.

Maybe we should just go ahead and copy paste, set up two separate shards?