@steve-the-edwards 2022-06-17
> 

Hmm. I'm loathe to copy-paste. But I think your point about multiple shards is good, because we could move the runtime choice from being in Test Util code to being in the executing environment. As in maybe we just use a system property for the sample.