@rjrjr 2022-06-17
> > 
> 
> Hmm. I'm loathe to copy-paste. But I think your point about multiple shards is good, because we could move the runtime choice from being in Test Util code to being in the executing environment. As in maybe we just use a system property for the sample tests.

Yeah, I just meant copy/paste of the github config -- I had assumed the tests were already configured from the environment and that GHA matrix itself was proving tricky. Certainly would not want to duplicate the test suites.